api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
tzdata==2025.1
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.9
APScheduler==3.10.4
# dev/test
pytest==8.3.3
httpx==0.27.2
ruff==0.6.9
black==24.10.0
